requirements.txt
asgiref==3.8.1
astroid==3.2.2
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.3
dill==0.3.8
Django==5.0.6
django-cors-headers==4.4.0
django-stubs==5.0.2
django-stubs-ext==5.0.2
djangorestframework==3.15.2
djangorestframework-stubs==3.15.0
exceptiongroup==1.2.1
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
nose==1.3.7
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pylint==3.2.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
requests==2.32.3
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
types-PyYAML==6.0.12.20240311
types-requests==2.32.0.20240602
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
